requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-telegram-bot==20.8
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
